refactor(types): split types_test.go into per-type test files
Split monolithic types_test.go into duration_test.go, time_test.go,
unix_test.go, and uuid_test.go for better organization and parity
with the one-file-per-type source structure.
